pyproject.toml
[project]
name = "rpgdapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "py-solc-x>=2.0.3",
    "python-dotenv>=1.0.1",
    "solc>=0.0.0a0",
    "uvicorn>=0.34.0",
    "web3>=7.8.0",
]
